Moved JavaVersion to an internal package
